TFY-4 Made the tests in the System Tests folder always use EC2 and moved all other tests to the EC2 Unit Tests folder.
