+++ ec2/tests/test_ec2_utils.py
@@ -17,6 +17,7 @@
 # Third Party Imports
 import testtools
 from moto import mock_ec2
+from boto.ec2 import EC2Connection
 
 # Cloudify Imports is imported and used in operations
 from ec2 import utils
@@ -91,3 +92,210 @@ class TestUtils(testtools.TestCase):
 
         self.assertEqual('agents', provider_context['agents_keypair'])
         self.assertEqual('agents', provider_context['agents_security_group'])
+
+    @mock_ec2
+    def test_utils_get_resource_id(self):
+
+        ctx = self.mock_ctx(
+            'test_utils_get_resource_id')
+        current_ctx.set(ctx=ctx)
+
+        ctx.node.properties['resource_id'] = \
+            'test_utils_get_resource_id'
+
+        resource_id = utils.get_resource_id()
+
+        self.assertEquals(
+            'test_utils_get_resource_id', resource_id)
+
+    @mock_ec2
+    def test_utils_get_resource_id_dynamic(self):
+
+        ctx = self.mock_ctx(
+            'test_utils_get_resource_id')
+        current_ctx.set(ctx=ctx)
+
+        ctx.node.properties['resource_id'] = ''
+
+        resource_id = utils.get_resource_id()
+
+        self.assertEquals('None-test_utils_get_resource_id', resource_id)
+
+    @mock_ec2
+    def test_utils_get_resource_id_from_key_path(self):
+
+        ctx = self.mock_ctx(
+            'test_utils_get_resource_id_from_key_path')
+        current_ctx.set(ctx=ctx)
+
+        ctx.node.properties['private_key_path'] = \
+            '~/.ssh/test_utils_get_resource_id_from_key_path.pem'
+
+        resource_id = utils.get_resource_id()
+
+        self.assertEquals(
+            'test_utils_get_resource_id_from_key_path', resource_id)
+
+    @mock_ec2
+    def test_utils_validate_node_properties_missing_key(self):
+        ctx = self.mock_ctx(
+            'test_utils_validate_node_properties_missing_key')
+        current_ctx.set(ctx=ctx)
+
+        ex = self.assertRaises(
+            NonRecoverableError, utils.validate_node_property,
+            'missing_key',
+            ctx.node.properties)
+
+        self.assertIn(
+            'missing_key is a required input. Unable to create.',
+            ex.message)
+
+    @mock_ec2
+    def test_utils_log_available_resources(self):
+
+        ctx = self.mock_ctx(
+            'test_utils_log_available_resources')
+        current_ctx.set(ctx=ctx)
+
+        client = EC2Connection()
+
+        key_pairs = client.get_all_key_pairs()
+
+        utils.log_available_resources(key_pairs)
+
+    @mock_ec2
+    def test_utils_get_external_resource_id_or_raise_no_id(self):
+
+        ctx = self.mock_ctx(
+            'test_utils_get_external_resource_id_or_raise_no_id')
+        current_ctx.set(ctx=ctx)
+
+        ctx.instance.runtime_properties['prop'] = None
+
+        ex = self.assertRaises(
+            NonRecoverableError,
+            utils.get_external_resource_id_or_raise,
+            'test_operation', ctx.instance)
+
+        self.assertIn(
+            'Cannot test_operation because {0} is not assigned'
+            .format(constants.EXTERNAL_RESOURCE_ID),
+            ex.message)
+
+    @mock_ec2
+    def test_utils_get_external_resource_id_or_raise(self):
+
+        ctx = self.mock_ctx(
+            'test_utils_get_external_resource_id_or_raise')
+        current_ctx.set(ctx=ctx)
+
+        ctx.instance.runtime_properties[constants.EXTERNAL_RESOURCE_ID] = \
+            'test_utils_get_external_resource_id_or_raise'
+
+        output = utils.get_external_resource_id_or_raise(
+            'test_operation', ctx.instance)
+
+        self.assertEquals(
+            'test_utils_get_external_resource_id_or_raise', output)
+
+    @mock_ec2
+    def test_utils_set_external_resource_id_cloudify(self):
+
+        ctx = self.mock_ctx(
+            'test_utils_set_external_resource_id_cloudify')
+        current_ctx.set(ctx=ctx)
+
+        utils.set_external_resource_id(
+            'id-value',
+            ctx.instance,
+            external=False)
+
+        self.assertEquals(
+            'id-value',
+            ctx.instance.runtime_properties[constants.EXTERNAL_RESOURCE_ID])
+
+    @mock_ec2
+    def test_utils_set_external_resource_id_external(self):
+
+        ctx = self.mock_ctx(
+            'test_utils_set_external_resource_id_external')
+        current_ctx.set(ctx=ctx)
+
+        utils.set_external_resource_id(
+            'id-value',
+            ctx.instance)
+
+        self.assertEquals(
+            'id-value',
+            ctx.instance.runtime_properties[constants.EXTERNAL_RESOURCE_ID])
+
+    @mock_ec2
+    def test_utils_unassign_runtime_property_from_resource(self):
+
+        ctx = self.mock_ctx(
+            'test_utils_unassign_runtime_property_from_resource')
+        current_ctx.set(ctx=ctx)
+
+        ctx.instance.runtime_properties[constants.EXTERNAL_RESOURCE_ID] = \
+            'test_utils_unassign_runtime_property_from_resource'
+
+        utils.unassign_runtime_property_from_resource(
+            constants.EXTERNAL_RESOURCE_ID,
+            ctx.instance)
+
+        self.assertNotIn(
+            constants.EXTERNAL_RESOURCE_ID,
+            ctx.instance.runtime_properties)
+
+    @mock_ec2
+    def test_utils_use_external_resource_not_external(self):
+
+        ctx = self.mock_ctx(
+            'test_utils_use_external_resource_not_external')
+        current_ctx.set(ctx=ctx)
+
+        self.assertEquals(
+            False,
+            utils.use_external_resource(ctx.node.properties))
+
+    @mock_ec2
+    def test_utils_use_external_resource_external(self):
+
+        ctx = self.mock_ctx(
+            'test_utils_use_external_resource_external')
+        current_ctx.set(ctx=ctx)
+
+        ctx.node.properties['use_external_resource'] = True
+        ctx.node.properties['resource_id'] = \
+            'test_utils_use_external_resource_external'
+
+        self.assertEquals(
+            True,
+            utils.use_external_resource(ctx.node.properties))
+
+    @mock_ec2
+    def test_get_target_external_resource_ids(self):
+
+        ctx = self.mock_ctx(
+            'get_target_external_resource_ids')
+        current_ctx.set(ctx=ctx)
+
+        output = utils.get_target_external_resource_ids(
+            'instance_connected_to_keypair',
+            ctx.instance)
+
+        self.assertEquals(0, len(output))
+
+    @mock_ec2
+    def test_get_target_external_resource_ids_no_attr(self):
+
+        ctx = self.mock_ctx(
+            'get_target_external_resource_ids')
+        current_ctx.set(ctx=ctx)
+
+        output = utils.get_target_external_resource_ids(
+            'instance_connected_to_keypair',
+            ctx.instance)
+
+        self.assertEquals(0, len(output))
