+++ system_tests/local/test_plugin.py
@@ -13,12 +13,8 @@
 #    * See the License for the specific language governing permissions and
 #    * limitations under the License.
 
-# Built-in Imports
-import os
-
 # Cloudify Imports
 from ec2 import (
-    utils,
     securitygroup,
     keypair,
     elasticip,
@@ -153,204 +149,6 @@ class TestWorkflowClean(EC2LocalTestUtils):
             str(instance_list[0].id))
 
 
-class EC2UtilsUnitTests(EC2LocalTestUtils):
-
-    def test_utils_get_resource_id(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_utils_get_resource_id')
-        current_ctx.set(ctx=ctx)
-
-        ctx.node.properties['resource_id'] = \
-            'test_utils_get_resource_id'
-
-        resource_id = utils.get_resource_id()
-
-        self.assertEquals(
-            'test_utils_get_resource_id', resource_id)
-
-    def test_utils_get_resource_id_dynamic(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_utils_get_resource_id')
-        current_ctx.set(ctx=ctx)
-
-        ctx.node.properties['resource_id'] = ''
-
-        resource_id = utils.get_resource_id()
-
-        self.assertEquals('None-test_utils_get_resource_id', resource_id)
-
-    def test_utils_get_resource_id_from_key_path(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_utils_get_resource_id_from_key_path')
-        current_ctx.set(ctx=ctx)
-
-        ctx.node.properties['private_key_path'] = \
-            '~/.ssh/test_utils_get_resource_id_from_key_path.pem'
-
-        resource_id = utils.get_resource_id()
-
-        self.assertEquals(
-            'test_utils_get_resource_id_from_key_path', resource_id)
-
-    def test_utils_validate_node_properties_missing_key(self):
-        ctx = self.mock_cloudify_context(
-            'test_utils_validate_node_properties_missing_key')
-        current_ctx.set(ctx=ctx)
-
-        ex = self.assertRaises(
-            NonRecoverableError, utils.validate_node_property,
-            'missing_key',
-            ctx.node.properties)
-
-        self.assertIn(
-            'missing_key is a required input. Unable to create.',
-            ex.message)
-
-    def test_utils_log_available_resources(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_utils_log_available_resources')
-        current_ctx.set(ctx=ctx)
-
-        client = self._get_ec2_client()
-
-        key_pairs = client.get_all_key_pairs()
-
-        utils.log_available_resources(key_pairs)
-
-    def test_utils_get_external_resource_id_or_raise_no_id(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_utils_get_external_resource_id_or_raise_no_id')
-        current_ctx.set(ctx=ctx)
-
-        ctx.instance.runtime_properties['prop'] = None
-
-        ex = self.assertRaises(
-            NonRecoverableError,
-            utils.get_external_resource_id_or_raise,
-            'test_operation', ctx.instance)
-
-        self.assertIn(
-            'Cannot test_operation because {0} is not assigned'
-            .format(EXTERNAL_RESOURCE_ID),
-            ex.message)
-
-    def test_utils_get_external_resource_id_or_raise(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_utils_get_external_resource_id_or_raise')
-        current_ctx.set(ctx=ctx)
-
-        ctx.instance.runtime_properties[EXTERNAL_RESOURCE_ID] = \
-            'test_utils_get_external_resource_id_or_raise'
-
-        output = utils.get_external_resource_id_or_raise(
-            'test_operation', ctx.instance)
-
-        self.assertEquals(
-            'test_utils_get_external_resource_id_or_raise', output)
-
-    def test_utils_set_external_resource_id_cloudify(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_utils_set_external_resource_id_cloudify')
-        current_ctx.set(ctx=ctx)
-
-        utils.set_external_resource_id(
-            'id-value',
-            ctx.instance,
-            external=False)
-
-        self.assertEquals(
-            'id-value',
-            ctx.instance.runtime_properties[EXTERNAL_RESOURCE_ID])
-
-    def test_utils_set_external_resource_id_external(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_utils_set_external_resource_id_external')
-        current_ctx.set(ctx=ctx)
-
-        utils.set_external_resource_id(
-            'id-value',
-            ctx.instance)
-
-        self.assertEquals(
-            'id-value',
-            ctx.instance.runtime_properties[EXTERNAL_RESOURCE_ID])
-
-    def test_utils_unassign_runtime_property_from_resource(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_utils_unassign_runtime_property_from_resource')
-        current_ctx.set(ctx=ctx)
-
-        ctx.instance.runtime_properties[EXTERNAL_RESOURCE_ID] = \
-            'test_utils_unassign_runtime_property_from_resource'
-
-        utils.unassign_runtime_property_from_resource(
-            EXTERNAL_RESOURCE_ID,
-            ctx.instance)
-
-        self.assertNotIn(
-            EXTERNAL_RESOURCE_ID,
-            ctx.instance.runtime_properties)
-
-    def test_utils_use_external_resource_not_external(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_utils_use_external_resource_not_external')
-        current_ctx.set(ctx=ctx)
-
-        self.assertEquals(
-            False,
-            utils.use_external_resource(ctx.node.properties))
-
-    def test_utils_use_external_resource_external(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_utils_use_external_resource_external')
-        current_ctx.set(ctx=ctx)
-
-        ctx.node.properties['use_external_resource'] = True
-        ctx.node.properties['resource_id'] = \
-            'test_utils_use_external_resource_external'
-
-        self.assertEquals(
-            True,
-            utils.use_external_resource(ctx.node.properties))
-
-    def test_get_target_external_resource_ids(self):
-
-        ctx = self.mock_cloudify_context(
-            'get_target_external_resource_ids')
-        current_ctx.set(ctx=ctx)
-
-        output = utils.get_target_external_resource_ids(
-            'instance_connected_to_keypair',
-            ctx.instance)
-
-        self.assertEquals(0, len(output))
-
-    def test_get_target_external_resource_ids_no_attr(self):
-
-        ctx = self.mock_cloudify_context(
-            'get_target_external_resource_ids')
-        current_ctx.set(ctx=ctx)
-
-        delattr(ctx.instance, 'relationships')
-
-        output = utils.get_target_external_resource_ids(
-            'instance_connected_to_keypair',
-            ctx.instance)
-
-        self.assertEquals(0, len(output))
-
-
 class EC2SecurityGroupUnitTests(EC2LocalTestUtils):
 
     def test_get_all_security_groups(self):
@@ -389,10 +187,10 @@ class EC2SecurityGroupUnitTests(EC2LocalTestUtils):
         group = client.create_security_group(
             'test_get_security_group_from_name',
             'some description')
+        self.addCleanup(group.delete)
         group_from_plugin = securitygroup._get_security_group_from_id(
             group_id=group.id)
         self.assertEqual(group.name, group_from_plugin.name)
-        group.delete()
 
     def test_get_security_group_from_id(self):
 
@@ -404,10 +202,10 @@ class EC2SecurityGroupUnitTests(EC2LocalTestUtils):
         group = client.create_security_group(
             'test_get_security_group_from_id',
             'some description')
+        self.addCleanup(group.delete)
         group_from_plugin = securitygroup._get_security_group_from_name(
             group_name=group.name)
         self.assertEqual(group.id, group_from_plugin.id)
-        group.delete()
 
     def test_get_security_group_from_name_but_really_id(self):
 
@@ -419,10 +217,10 @@ class EC2SecurityGroupUnitTests(EC2LocalTestUtils):
         group = client.create_security_group(
             'test_get_security_group_from_name_but_really_id',
             'some description')
+        self.addCleanup(group.delete)
         group_from_plugin = securitygroup._get_security_group_from_name(
             group_name=group.id)
         self.assertEqual(group.name, group_from_plugin.name)
-        group.delete()
 
     def test_get_security_group_from_id_but_really_name(self):
 
@@ -434,10 +232,10 @@ class EC2SecurityGroupUnitTests(EC2LocalTestUtils):
         group = client.create_security_group(
             'test_get_security_group_from_id_but_really_name',
             'some description')
+        self.addCleanup(group.delete)
         group_from_plugin = securitygroup._get_security_group_from_id(
             group_id=group.name)
         self.assertEqual(group.id, group_from_plugin.id)
-        group.delete()
 
     def test_delete_external_securitygroup_external(self):
 
@@ -454,17 +252,6 @@ class EC2SecurityGroupUnitTests(EC2LocalTestUtils):
         self.assertNotIn(
             EXTERNAL_RESOURCE_ID, ctx.instance.runtime_properties)
 
-    def test_delete_external_securitygroup_not_external(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_delete_external_securitygroup_not_external')
-        current_ctx.set(ctx=ctx)
-
-        ctx.node.properties['use_external_resource'] = False
-
-        output = securitygroup._delete_external_securitygroup()
-        self.assertEqual(False, output)
-
     def test_create_external_securitygroup_external(self):
 
         ctx = self.mock_cloudify_context(
@@ -475,7 +262,7 @@ class EC2SecurityGroupUnitTests(EC2LocalTestUtils):
         group = client.create_security_group(
             'test_create_external_securitygroup_external',
             'some description')
-
+        self.addCleanup(group.delete)
         ctx.node.properties['use_external_resource'] = True
         ctx.node.properties['resource_id'] = group.id
 
@@ -484,7 +271,6 @@ class EC2SecurityGroupUnitTests(EC2LocalTestUtils):
         self.assertEqual(
             ctx.instance.runtime_properties[EXTERNAL_RESOURCE_ID],
             group.id)
-        group.delete()
 
     def test_create_external_securitygroup_external_bad_id(self):
 
@@ -495,22 +281,11 @@ class EC2SecurityGroupUnitTests(EC2LocalTestUtils):
         ctx.node.properties['use_external_resource'] = True
         ctx.node.properties['resource_id'] = 'sg-73cd3f1e'
 
-        ex = self.assertRaises(
-            NonRecoverableError,
-            securitygroup._create_external_securitygroup,
-            'sg-73cd3f1e')
-        self.assertIn('security group does not exist', ex.message)
-
-    def test_create_external_securitygroup_not_external(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_create_external_securitygroup_not_external')
-        current_ctx.set(ctx=ctx)
-
-        ctx.node.properties['use_external_resource'] = False
-
-        output = securitygroup._delete_external_securitygroup()
-        self.assertEqual(False, output)
+        with self.assertRaisesRegexp(
+                NonRecoverableError,
+                'security group does not exist'):
+            securitygroup._create_external_securitygroup(
+                'sg-73cd3f1e')
 
     def test_create_group_rules_ruleset(self):
 
@@ -531,6 +306,7 @@ class EC2SecurityGroupUnitTests(EC2LocalTestUtils):
         group = client.create_security_group(
             'test_create_group_rules',
             'some description')
+        self.addCleanup(group.delete)
         securitygroup._create_group_rules(group)
         groups_from_test = \
             client.get_all_security_groups(groupnames=[group.name])
@@ -539,109 +315,20 @@ class EC2SecurityGroupUnitTests(EC2LocalTestUtils):
             str(groups_from_test[0].rules[0]),
             'IPPermissions:tcp(22-22)'
         )
-        group.delete()
 
     def test_delete_security_group_bad_group(self):
         ctx = self.mock_cloudify_context(
             'test_delete_security_group_bad_group')
         current_ctx.set(ctx=ctx)
 
-        ex = self.assertRaises(
-            NonRecoverableError,
-            securitygroup._delete_security_group, 'sg-73cd3f1e')
+        with self.assertRaisesRegexp(
+                NonRecoverableError,
+                'does not exist in the account'):
 
-        self.assertIn(
-            'because the group does not exist in the account', ex.message)
+            securitygroup._delete_security_group('sg-73cd3f1e')
 
 
 class EC2KeyPairUnitTests(EC2LocalTestUtils):
-
-    def test_search_for_key_file_no_file(self):
-
-        output = keypair._search_for_key_file(
-            '~/.ssh/test_search_for_key_file.pem')
-
-        self.assertEquals(
-            False,
-            output
-        )
-
-    def test_get_path_to_key_folder_no_private_key_path(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_get_path_to_key_folder_no_private_key_path')
-        current_ctx.set(ctx=ctx)
-        ex = self.assertRaises(
-            NonRecoverableError,
-            keypair._get_path_to_key_file)
-
-        self.assertIn(
-            'private_key_path not set',
-            ex.message
-        )
-
-    def test_get_path_to_key_folder(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_get_path_to_key_folder')
-        current_ctx.set(ctx=ctx)
-        ctx.node.properties['private_key_path'] = \
-            '~/.ssh/test_get_path_to_key_folder.pem'
-
-        full_key_path = os.path.expanduser(
-            ctx.node.properties['private_key_path']
-        )
-
-        key_directory, key_filename = os.path.split(full_key_path)
-
-        output = keypair._get_path_to_key_folder()
-
-        self.assertEqual(key_directory, output)
-
-    def test_get_path_to_key_file_no_private_key_path(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_get_path_to_key_folder_no_private_key_path')
-        current_ctx.set(ctx=ctx)
-        ex = self.assertRaises(
-            NonRecoverableError,
-            keypair._get_path_to_key_file)
-
-        self.assertIn(
-            'private_key_path not set',
-            ex.message
-        )
-
-    def test_get_path_to_key_file(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_get_path_to_key_folder')
-        current_ctx.set(ctx=ctx)
-        ctx.node.properties['private_key_path'] = \
-            '~/.ssh/test_get_path_to_key_folder.pem'
-
-        full_key_path = os.path.expanduser(
-            ctx.node.properties['private_key_path']
-        )
-
-        output = keypair._get_path_to_key_file()
-
-        self.assertEqual(full_key_path, output)
-
-    def test_get_key_pair_by_id_no_kp(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_get_key_pair_by_id_no_kp')
-        current_ctx.set(ctx=ctx)
-
-        ex = self.assertRaises(
-            NonRecoverableError,
-            keypair._get_key_pair_by_id,
-            'test_get_key_pair_by_id_no_kp')
-
-        self.assertIn(
-            'InvalidKeyPair.NotFound',
-            ex.message)
 
     def test_get_key_pair_by_id(self):
 
@@ -652,23 +339,12 @@ class EC2KeyPairUnitTests(EC2LocalTestUtils):
         client = self._get_ec2_client()
         kp = client.create_key_pair(
             'test_get_key_pair_by_id')
-
+        self.addCleanup(kp.delete)
         output = keypair._get_key_pair_by_id(kp.name)
         self.assertEqual(kp.name, output.name)
-        kp.delete()
 
 
 class EC2ElasticIPUnitTests(EC2LocalTestUtils):
-
-    def test_get_all_addresses_bad_address(self):
-
-        ctx = self.mock_relationship_context(
-            'test_get_address_by_id')
-        current_ctx.set(ctx=ctx)
-
-        output = elasticip._get_all_addresses('127.0.0.1')
-
-        self.assertIsNone(output)
 
     def test_get_address_object_by_id(self):
 
@@ -680,9 +356,9 @@ class EC2ElasticIPUnitTests(EC2LocalTestUtils):
         address = client.allocate_address()
         address_object = \
             elasticip._get_address_object_by_id(address.public_ip)
+        self.addCleanup(address_object.delete)
         self.assertEqual(
             address.public_ip, address_object.public_ip)
-        address_object.delete()
 
     def test_get_address_by_id(self):
 
@@ -692,21 +368,9 @@ class EC2ElasticIPUnitTests(EC2LocalTestUtils):
 
         client = self._get_ec2_client()
         address_object = client.allocate_address()
+        self.addCleanup(address_object.delete)
         address = elasticip._get_address_by_id(address_object.public_ip)
         self.assertEqual(address, address_object.public_ip)
-        address_object.delete()
-
-    def test_disassociate_external_elasticip_or_instance(self):
-
-        ctx = self.mock_relationship_context(
-            'test_disassociate_external_elasticip_or_instance')
-        current_ctx.set(ctx=ctx)
-        ctx.source.node.properties['use_external_resource'] = False
-
-        output = \
-            elasticip._disassociate_external_elasticip_or_instance()
-
-        self.assertEqual(False, output)
 
     def test_disassociate_external_elasticip_or_instance_external(self):
 
@@ -726,19 +390,6 @@ class EC2ElasticIPUnitTests(EC2LocalTestUtils):
             'public_ip_address',
             ctx.source.instance.runtime_properties)
 
-    def test_associate_external_elasticip_or_instance(self):
-
-        ctx = self.mock_relationship_context(
-            'test_associate_external_elasticip_or_instance')
-        current_ctx.set(ctx=ctx)
-        ctx.source.node.properties['use_external_resource'] = False
-
-        output = \
-            elasticip._associate_external_elasticip_or_instance(
-                '127.0.0.1')
-
-        self.assertEqual(False, output)
-
     def test_associate_external_elasticip_or_instance_external(self):
 
         ctx = self.mock_relationship_context(
@@ -746,6 +397,7 @@ class EC2ElasticIPUnitTests(EC2LocalTestUtils):
         current_ctx.set(ctx=ctx)
         client = self._get_ec2_client()
         address_object = client.allocate_address()
+        self.addCleanup(address_object.delete)
 
         ctx.source.node.properties['use_external_resource'] = True
         ctx.target.node.properties['use_external_resource'] = True
@@ -764,19 +416,6 @@ class EC2ElasticIPUnitTests(EC2LocalTestUtils):
         self.assertEqual(
             address_object.public_ip,
             ctx.source.instance.runtime_properties['public_ip_address'])
-        address_object.delete()
-
-    def test_release_external_elasticip(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_release_external_elasticip')
-        current_ctx.set(ctx=ctx)
-        ctx.node.properties['use_external_resource'] = False
-
-        output = \
-            elasticip._release_external_elasticip()
-
-        self.assertEqual(False, output)
 
     def test_release_external_elasticip_external(self):
 
@@ -795,18 +434,6 @@ class EC2ElasticIPUnitTests(EC2LocalTestUtils):
             EXTERNAL_RESOURCE_ID,
             ctx.instance.runtime_properties)
 
-    def test_allocate_external_elasticip(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_allocate_external_elasticip')
-        current_ctx.set(ctx=ctx)
-        ctx.node.properties['use_external_resource'] = False
-
-        output = \
-            elasticip._allocate_external_elasticip()
-
-        self.assertEqual(False, output)
-
     def test_allocate_external_elasticip_external(self):
 
         ctx = self.mock_cloudify_context(
@@ -814,7 +441,7 @@ class EC2ElasticIPUnitTests(EC2LocalTestUtils):
         current_ctx.set(ctx=ctx)
         client = self._get_ec2_client()
         address_object = client.allocate_address()
-
+        self.addCleanup(address_object.delete)
         ctx.node.properties['use_external_resource'] = True
         ctx.node.properties['resource_id'] = address_object.public_ip
 
@@ -829,8 +456,6 @@ class EC2ElasticIPUnitTests(EC2LocalTestUtils):
             address_object.public_ip,
             ctx.instance.runtime_properties[EXTERNAL_RESOURCE_ID])
 
-        address_object.delete()
-
     def test_allocate_external_elasticip_external_bad_id(self):
 
         ctx = self.mock_cloudify_context(
@@ -839,13 +464,10 @@ class EC2ElasticIPUnitTests(EC2LocalTestUtils):
         ctx.node.properties['use_external_resource'] = True
         ctx.node.properties['resource_id'] = '127.0.0.1'
 
-        ex = self.assertRaises(
-            NonRecoverableError,
-            elasticip._allocate_external_elasticip)
-
-        self.assertIn(
-            'the given elasticip does not exist in the account',
-            ex.message)
+        with self.assertRaisesRegexp(
+                NonRecoverableError,
+                'elasticip does not exist in the account'):
+            elasticip._allocate_external_elasticip()
 
 
 class EC2InstanceUnitTests(EC2LocalTestUtils):
@@ -857,10 +479,10 @@ class EC2InstanceUnitTests(EC2LocalTestUtils):
 
         image_id = 'ami-65b95565'
 
-        ex = self.assertRaises(
-            NonRecoverableError, instance._get_image, image_id)
-
-        self.assertIn('InvalidAMIID.NotFound', ex.message)
+        with self.assertRaisesRegexp(
+                NonRecoverableError,
+                'InvalidAMIID.NotFound'):
+            instance._get_image(image_id)
 
     def test_instance_get_image_id(self):
 
@@ -882,34 +504,7 @@ class EC2InstanceUnitTests(EC2LocalTestUtils):
         ctx.node.properties['use_external_resource'] = True
         ctx.node.properties['resource_id'] = 'i-00z0zz0z'
 
-        ex = self.assertRaises(
-            NonRecoverableError, instance._create_external_instance)
-
-        self.assertIn('is not in this account', ex.message)
-
-    def test_get_instance_keypair(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_get_instance_keypair')
-        current_ctx.set(ctx=ctx)
-
-        provider_variables = {
-            'agents_keypair': '',
-            'agents_security_group': ''
-        }
-        output = instance._get_instance_keypair(provider_variables)
-        self.assertEqual(None, output)
-
-    def test_get_instance_parameters(self):
-
-        ctx = self.mock_cloudify_context(
-            'test_get_instance_parameters')
-        current_ctx.set(ctx=ctx)
-        ctx.node.properties['image_id'] = 'abc'
-        ctx.node.properties['instance_type'] = 'efg'
-        ctx.node.properties['parameters']['image_id'] = 'abcd'
-        ctx.node.properties['parameters']['key_name'] = 'xyz'
-        parameters = instance._get_instance_parameters()
-        self.assertIn('abcd', parameters['image_id'])
-        self.assertIn('xyz', parameters['key_name'])
-        self.assertIn('efg', parameters['instance_type'])
+        with self.assertRaisesRegexp(
+                NonRecoverableError,
+                'not in this account'):
+            instance._create_external_instance()
